+++ CHANGELOG.md
@@ -5,12 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.0.8 - Unreleased]
+## [1.0.0.8]
 ### Added
 - Framingassistant
 - Switchpage
 - portuguese support
 - Manuell mount control
+- Guider notes
   
 ### Changed
 - Filter wheel cannot be connected if it is manual
@@ -27,6 +28,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Weather modal
 - Android Update
 - Subnav
+- Guider state management
+- Guider RA / DEC values
 
 ## [1.0.0.7] - 2025-01-27
 ### Added